test[staging-provider] Workspace and WorkspaceWithCredentials tests
